# docs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "MoneyKeeper API",
    "description": "This is a simple MoneyKeeper software.",
    "license": {
      "name": "Apache 2.0",
      "url": "http://alnat.dev"
    },
    "version": "1"
  },
  "servers": [
    {
      "url": "http://localhost:8080/",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/icon/{iconID}": {
      "get": {
        "tags": [
          "Icon API"
        ],
        "summary": "Получение иконки по идентификатору",
        "operationId": "getIconByID",
        "parameters": [
          {
            "name": "iconID",
            "in": "path",
            "description": "Идентификатор иконки",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Иконки с таким идентификатором не найдено",
            "content": {}
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Icon"
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "delete": {
        "tags": [
          "Icon API"
        ],
        "summary": "Удаление иконки",
        "operationId": "deleteIcon",
        "parameters": [
          {
            "name": "iconID",
            "in": "path",
            "description": "Идентификатор иконки",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "example": 1
          }
        ],
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "400": {
            "description": "Запрос не корректен",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав для запроса"
          },
          "404": {
            "description": "Категории с таким идентификатором не найдено",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "200": {
            "description": "Удаление успешно выполнено"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
    "/api/icon/": {
      "get": {
        "tags": [
          "Icon API"
        ],
        "summary": "Получение списка иконок",
        "operationId": "getIconList",
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "200": {
            "description": "Запрос успешно выполнен",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Icon"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "put": {
        "tags": [
          "Icon API"
        ],
        "summary": "Обновление иконки",
        "operationId": "updatedIcon",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Icon"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "201": {
            "description": "Иконка успешно обновлена"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      },
      "post": {
        "tags": [
          "Icon API"
        ],
        "summary": "Загрузка новой иконки",
        "operationId": "addIcon",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "500": {
            "description": "Ошибка при обработки запроса",
            "content": {}
          },
          "400": {
            "description": "Ошибка в запросе",
            "content": {}
          },
          "403": {
            "description": "Недостаточно прав для запроса",
            "content": {}
          },
          "404": {
            "description": "Not Found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            }
          },
          "201": {
            "description": "Иконка успешно создана"
          },
          "401": {
            "description": "Запрос не авторизован"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Описание ошибки"
          }
        }
      },
...
      "Icon": {
        "required": [
          "data"
        ],
        "type": "object",
        "properties": {
          "created": {
            "type": "string",
            "description": "Дата создания записи",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "description": "Дата обновления записи",
            "format": "date-time"
          },
          "iconID": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "byte"
            }
          }
        },
        "description": "Измененная иконка"
      },
...
    }
  }
}